30_code/pathway_rag/requirements.txt
pathway[all]
fastapi
uvicorn
openai
python-dotenv
gunicorn